power_automate/azure_function/requirements.txt
azure-functions
pdfplumber
